+++ nodebackend/src/routes/loginRoute.ts
@@ -8,7 +8,8 @@ import { updateLastLogin, getLastLogin } from '../utils/useLoginsModule';
 import jwt from 'jsonwebtoken';
 import { isSecureCookie, jwtTokenExpiry } from '../envSwitcher';
 import { getJwtAccessTokenSecret } from '../configs';
-import { initializeEncryptionKey, encrypt } from '../utils/enyryptDecrypt'; 
+import { initializeEncryptionKey, encrypt } from '../utils/enyryptDecrypt';
+import generateUniqueId from '../utils/generateUniqueId';
 
 const router = express.Router();
 
@@ -42,6 +43,8 @@ router.post('/', async (req: express.Request, res: express.Response) => {
             return res.status(401).json({ status: 'error', message: 'incorrectUserOrPass' });
         }
 
+        const deviceId = generateUniqueId();
+
         const jwtSecret = await getJwtAccessTokenSecret();
         const expiresIn = jwtTokenExpiry;
 
@@ -51,7 +54,7 @@ router.post('/', async (req: express.Request, res: express.Response) => {
 
         const redis = await connectToRedis();
         const refreshTokenData = JSON.stringify({ refreshToken, userRole });
-        await redis.set(`refreshToken:${username}`, refreshTokenData, 'EX', 30 * 24 * 60 * 60);
+        await redis.set(`refreshToken:${username}_${deviceId}`, refreshTokenData, 'EX', 30 * 24 * 60 * 60);
 
         res.cookie('refreshToken', refreshToken, {
             httpOnly: true,
@@ -76,7 +79,7 @@ router.post('/', async (req: express.Request, res: express.Response) => {
         await updateLastLogin(username);
         logger.info(`User ${username} logged in successfully from IP ${clientIp}`);
 
-        res.status(200).json({ status: 'success', accessToken, previousLastLogin: previousLastLogin, message: 'loggedIn' });
+        res.status(200).json({ status: 'success', accessToken, deviceId, previousLastLogin: previousLastLogin, message: 'loggedIn' });
     } catch (error) {
         if (error instanceof Error) {
             logger.error(`Error during user login for username: ${username} from IP ${clientIp} - ${error.message}`);